requirements.txt
streamlit==1.33.0
pydantic>=2.7
supabase-py>=2.0
SQLAlchemy>=2.0
asyncpg>=0.29
python-dotenv>=1.0
streamlit-authenticator>=0.3
alembic>=1.13
pytest>=8.1
ruff>=0.4
